requirements.txt
pandas
geopandas
pickledb==0.9.2
scikit-Learn
qwikidata
PyYAML
IPython
geopandas